openpilot v0.8.5

+++ cereal/log.capnp
@@ -541,7 +541,7 @@ struct ControlsState @0x97ff69c53601abf1 {
   limitSpeedCameraDist @63 :Float32;
   steerRatio @64 :Float32;
   longPlanSource  @65 :UInt8;
-  mapSign @66 :UInt8;
+  mapSign @66 :Float32;
 
   lateralControlState :union {
     indiState @52 :LateralINDIState;
@@ -798,7 +798,7 @@ struct LongitudinalPlan @0xe00b5b3eba12876c {
   status2 @38 :Bool;
   targetSpeedCamera @39 :Float32;
   targetSpeedCameraDist @40 :Float32;
-  mapSign @41 :UInt8;
+  mapSign @41 :Float32;
 
   enum LongitudinalPlanSource {
     cruise @0;
